+++ resources/views/components/schema-node-inspector.blade.php
@@ -1,0 +1,12 @@
+@props(['defaultUi' => true])
+
+<aside
+    {{ $attributes->class(['flow-schema-node-inspector']) }}
+    x-schema-node-inspector
+>
+    @if($defaultUi && $slot->isEmpty())
+        <template x-schema-default-ui></template>
+    @else
+        {{ $slot }}
+    @endif
+</aside>